% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4Seas.helpers.R
\name{adjustPrevSeasonENSOplume}
\alias{adjustPrevSeasonENSOplume}
\title{Infer season of the previous observed time series}
\usage{
adjustPrevSeasonENSOplume(season, n.prev.months = 4)
}
\arguments{
\item{season}{Season of the observed reference used as reference for ACC and anomaly calculation}

\item{n.prev.months}{Number of previous months used before initialization. Default to 4, as usual in C3S plumes}
}
\value{
An integer vector, with the season definition
}
\description{
Infer the season of the previous observed time series in ENSO plume plots
}
\details{
interna helper used for metaclipR.ENSOplume in the QA4Seas prototype
}
\keyword{internal}
