% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4Seas.helpers.R
\name{checkModelType}
\alias{checkModelType}
\title{Check model type}
\usage{
checkModelType(model, model.type)
}
\arguments{
\item{model}{model string, as read from the parameters file}

\item{model.type}{Model type. Accepted values are \code{"hindcast"}, \code{"forecast"} and \code{"ref"}}
}
\value{
Acts silently. Stops the execution if the check is not fulfilled
}
\description{
Checks the consistency between the model string and the type assigned (e.g., to avoid
defining ERA-Interim as a seasonal forecast).
}
\keyword{internal}
