% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4Seas.helpers.R
\name{defineNmembers}
\alias{defineNmembers}
\title{Define number of members}
\usage{
defineNmembers(model, model.type, combination.method)
}
\arguments{
\item{model}{model string, as read from the parameters file}

\item{model.type}{Model type. Accepted values are \code{"hindcast"}, \code{"forecast"} and \code{"ref"},
the latter yielding \code{NULL} (i.e., no members)}

\item{combination.method}{Combination method as inficated in the parameter file}
}
\value{
Integer, number of members
}
\description{
Ascertain the number of members of the dataset from the parameter file
}
\keyword{internal}
