% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4Seas.helpers.R
\name{getNdays}
\alias{getNdays}
\title{Calculate the number of days of a given month}
\usage{
getNdays(d)
}
\arguments{
\item{d}{A date (character) in format YYYY-MM-DD...}
}
\value{
The number of days of the current month
}
\description{
Calculate the number of days of the current month (it takes into account leap years for februaries)
}
\keyword{internal}
