% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4Seas.helpers.R
\name{parseSeason}
\alias{parseSeason}
\title{Parse product forecast time}
\usage{
parseSeason(init, season, several.ok = TRUE)
}
\arguments{
\item{init}{Character string, of the form \code{"fcXXXX"}, as read from the parameter file}

\item{season}{Character string, as read from the \code{'forecast_time_intervals'} product parameter}

\item{several.ok}{Logical flag indicating whether the product is displaying several forecast
times in the same plot (e.g., and ENSO plume) or not (e.g., a skill map)}
}
\value{
A list, whose length is equal to the number of products to be generated
}
\description{
Extracts the season from the parameters read from the yml file
}
\keyword{internal}
