% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4Seas.DatasetSubset.R
\name{qa4seas.DatasetSubset}
\alias{qa4seas.DatasetSubset}
\title{METACLIP definition of DatasetSubsets for QA4Seas prototype outcomes}
\usage{
qa4seas.DatasetSubset(package, version, graph, var_code, init, par.list,
  RefSpatialExtent, model, model.type, season, prod.info,
  proj = "LatLon", resX = "1_deg", resY = "1_deg")
}
\arguments{
\item{package}{Software Package string}

\item{version}{Software package version string}

\item{graph}{Input graph (contains parent node).}

\item{var_code}{Character string with variable code}

\item{init}{Initialization time parameter. Passed by upper function}

\item{par.list}{Full parameter list}

\item{RefSpatialExtent}{A metaclipR object containing the spatial extent definition (HorizontalExtent-class)
\code{"ENSO3"}, \code{"ENSO3dot4"}, \code{"ENSO4"}, and \code{"GlobalExtent"}}

\item{model}{Model designation, as indicated in the paraneter file}

\item{model.type}{Character string. Possible choices are \code{"ref"},
\code{"hindcast"} or \code{"forecast"}.}

\item{season}{Season vector (passed by upper function)}

\item{prod.info}{Product parameters, previously read}

\item{proj}{Projection. Unused, default to \code{"LatLon"}}

\item{resX}{Character. Spatial resolution of X coordinates. Unused, default to \code{"1_deg"}}

\item{resY}{Same as \code{resX}, for Y-coordinates}
}
\description{
Uses information from the parameter file to extract the DatasetSubset-class definition
}
