% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4Seas.EnsoPlume.R
\name{qa4seas.EnsoPlume}
\alias{qa4seas.EnsoPlume}
\title{Directed metadata graph construction for ENSO plume graphical products}
\usage{
qa4seas.EnsoPlume(package = "QA4Seas-prototype", version = "1.0.1",
  fun = "QA4seas.py", multimodel = FALSE, arg.list = NULL,
  FcstMetadata, ObsMetadata = NULL, VerificationMetadata = NULL,
  title = "ENSO Plume")
}
\arguments{
\item{package}{package}

\item{version}{version}

\item{fun}{function name.}

\item{multimodel}{Logical. Is the forecast done with a multimodel ensemble dataset?. Default
to \code{FALSE} (This is needed to ascertain the subtypes of ForecastSystemUncertainty communicated
by the product, i.e. FSi/FSm).}

\item{arg.list}{Argument list}

\item{FcstMetadata}{The metadata structure containing the step describing the forecast data provenance}

\item{ObsMetadata}{The metadata structure containing the step describing the observation data provenance}

\item{VerificationMetadata}{The metadata structure containing the step describing the verification provenance.
Only needed when EQC info (ACC) is added to the plume}

\item{title}{Character string character with the plot title}

\item{multimodel}{Logical. Is it a multi-model product?}
}
\description{
Build a directed metadata graph describing a ENSO plume plot from the QA4Seas prototype
}
\concept{graphical.products}
\keyword{internal}
