% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4seas.metaclipR.R
\name{qa4seas.metaclipR}
\alias{qa4seas.metaclipR}
\title{QA4Seas prototype metaclip launcher}
\usage{
qa4seas.metaclipR(parameter.file, pkg = "QA4Seas-prototype",
  v = "1.0.1", out.dir)
}
\arguments{
\item{parameter.file}{Input yaml parameter file}

\item{pkg}{Software package. Default to \code{"QA4Seas-prototype"}, which has an individual description in the datasource ontology.}

\item{v}{Software package version character string. Default to \code{"1.0.1"}}

\item{out.dir}{Full path to the output directory the output json files are to be stored}
}
\description{
Given a prototype input yaml file, parses the parameters to yield json files with metadata description for
 the outcomes generated
}
