% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4Seas.helpers.R
\name{setAggrArgList}
\alias{setAggrArgList}
\title{Derive aggregation arg.list}
\usage{
setAggrArgList(product.type, var_code)
}
\arguments{
\item{product.type}{product type parameter string}

\item{var_code}{variable code string}
}
\value{
An argument list, as required by \code{\link[metaclipR]{metaclipR.Aggregation}}
}
\description{
Infer the aggregation parameters list given a variable and product type
}
\keyword{internal}
