% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa4Seas.helpers.R
\name{startModelGraph}
\alias{startModelGraph}
\title{Start a QA4Seas dataset graph}
\usage{
startModelGraph(pkg = "QA4Seas-prototype", v = "1.0.1",
  fun = "QA4Seas.py", par.list, model.type, RefSpatialExtent, var_code,
  prod.info, init, season, disable.command = TRUE)
}
\arguments{
\item{pkg}{pkg}

\item{v}{version}

\item{fun}{function}

\item{par.list}{Full list of parameters}

\item{model.type}{Character string specifying the type of model.}

\item{RefSpatialExtent}{Ref spatial extent}

\item{var_code}{individual var code}

\item{prod.info}{individual product info list}

\item{init}{individual init}

\item{season}{individual season}

\item{disable.command.}{Default to \code{TRUE}. Passed to \code{\link{qa4seas.DatasetSubset}}}
}
\value{
A metaclipR structure. The terminal node belong either to the specific subclass of Dataset-class
for model lists of length one, or ds:Ensemble-class, in case several models are included in
 the \code{forecasting_system} parameter.
}
\description{
Start a QA4Seas dataset graph fiev a dataset (possibly a multimodel)
}
\details{
The function internally handles the recursive listing in case of multimodel ensembles.
The specific inputs come from the ouer wrapping function
}
\keyword{internal}
